+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Other blocks are:
 
 - **BLS** - Ranges of subplots are supported in the definition of the curves.
 - **BLS** - Support for multiple spectra in a figure.
+- **VIZ** - Simple mechanism to add a panel label in an angle of a spectrum in `SpecLayout`.
 
 ### Fixed